requirements.txt
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
nltk>=3.7
rouge-score>=0.1.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
requests>=2.28.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
pyyaml>=6.0
sacrebleu>=2.3.0
gdown>=4.6.0
